session01/005bowClassification/requirements.txt
scikit-learn>=1.3
nltk>=3.8
joblib>=1.3
matplotlib>=3.7
pandas>=2.0
numpy>=1.25
PyYAML>=6.0
